Field dumps (E-field) TD, hard E-source

+++ FDTD/engine.cpp
@@ -62,7 +62,8 @@ bool Engine::IterateTS(unsigned int iterTS)
 		for (unsigned int n=0;n<Op->E_Ex_Count;++n)
 		{
 			exc_pos = (int)numTS - (int)Op->E_Ex_delay[n];
-			exc_pos*= (exc_pos>0 && exc_pos<Op->ExciteLength);
+			exc_pos*= (exc_pos>0 && exc_pos<(int)Op->ExciteLength);
+//			if (n==0) cerr << numTS << " => " << Op->ExciteSignal[exc_pos] << endl;
 			volt[0][Op->E_Ex_index[0][n]][Op->E_Ex_index[1][n]][Op->E_Ex_index[2][n]] += Op->E_Ex_amp[0][n]*Op->ExciteSignal[exc_pos];
 			volt[1][Op->E_Ex_index[0][n]][Op->E_Ex_index[1][n]][Op->E_Ex_index[2][n]] += Op->E_Ex_amp[1][n]*Op->ExciteSignal[exc_pos];
 			volt[2][Op->E_Ex_index[0][n]][Op->E_Ex_index[1][n]][Op->E_Ex_index[2][n]] += Op->E_Ex_amp[2][n]*Op->ExciteSignal[exc_pos];
